Crate pallet_tickets

This pallet frag performs logic related to FRAG Token

IMPORTANT NOTE: The term “lock” refers to the “effective transfer” of some ERC-20 FRAG tokens from Fragnova-owned FRAG Ethereum Smart Contract to the Clamor Blockchain.

The term “unlock” refers to the “effective transfer” of all the previously locked ERC-20 FRAG tokens from the Clamor Blockchain to the aforementioned Fragnova-owned FRAG Ethereum Smart Contract.

The term “staking” refers to the staking of the FRAG Token that is done in the Clamor Blockchain itself. It is different to the term “locking” which is defined above.

IMPORTANT: locking != staking

Re-exports

pub use pallet::*;

Modules

crypto

Based on the above KeyTypeId we need to generate a pallet-specific crypto type wrappers. We can use from supported crypto kinds (sr25519, ed25519 and ecdsa) and augment the types with this pallet-specific identifier.

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Structs

EthLock

Struct representing the details about the total amount of locked FRAG Token of a particular Ethereum Account in the Fragnova-owned Ethereum Smart Contract .

EthLockUpdate

Struct representing a recent confirmed (i.e with sufficient blockchain confirmations) log for the event Lock or Unlock of the FRAG token Smart Contract

Constants

KEY_TYPE

Defines application identifier for crypto keys of this module.

Traits

EthFragContract

Traits of the FRAG Token Smart Contract on the Ethereum Blockchain

WeightInfo

Weight functions needed for pallet_protos.