Trait pallet_tickets::pallet::Config

pub trait Config: Config + CreateSignedTransaction<Call<Self>> + Config {
    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type WeightInfo: WeightInfo;
    type EthChainId: Get<u64>;
    type EthConfirmations: Get<u64>;
    type EthFragContract: EthFragContract;
    type Threshold: Get<u64>;
    type AuthorityId: AppCrypto<Self::Public, Self::Signature>;
}

Configure the pallet by specifying the parameters and types on which it depends.

Required Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

Because this pallet emits events, it depends on the runtime’s definition of an event.

type WeightInfo: WeightInfo

type EthChainId: Get<u64>

type EthConfirmations: Get<u64>

The number of confirmations required to consider a transaction on the Ethereum Blockchain final (https://www.youtube.com/watch?v=gP5zcHD8tJU)

type EthFragContract: EthFragContract

type Threshold: Get<u64>

Number of votes needed to do something (问Gio)

type AuthorityId: AppCrypto<Self::Public, Self::Signature>

The identifier type for an offchain worker.